@@ -1,57 +1,61 @@
 {
-  "name": "notion-workspace",
-  "version": "0.0.0-beta",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "watch": "ng build --watch --configuration development",
-    "test": "ng test",
-    "lint": "ng lint",
-    "ngx-notion-cms:lint": "ng lint ngx-notion-cms",
-    "ngx-notion-cms:test": "ng test ngx-notion-cms --no-watch --no-progress --browsers=ChromeHadless",
-    "ngx-notion-cms:build": "ng build ngx-notion-cms",
-    "ngx-notion-cms:publish": "npm run ngx-notion-cms:build && npm publish dist/ngx-notion-cms/",
-    "serve:ssr:ngx-notion-testbed-app": "node dist/ngx-notion-testbed-app/server/server.mjs"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "^18.2.0",
-    "@angular/common": "^18.2.0",
-    "@angular/compiler": "^18.2.0",
-    "@angular/core": "^18.2.0",
-    "@angular/forms": "^18.2.0",
-    "@angular/platform-browser": "^18.2.0",
-    "@angular/platform-browser-dynamic": "^18.2.0",
-    "@angular/platform-server": "^18.2.0",
-    "@angular/router": "^18.2.0",
-    "@angular/ssr": "^18.2.5",
-    "express": "^4.18.2",
-    "ngx-highlightjs": "^12.0.0",
-    "rxjs": "~7.8.0",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.14.10"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "^18.2.5",
-    "@angular/cli": "^18.2.1",
-    "@angular/compiler-cli": "^18.2.0",
-    "@types/express": "^4.17.17",
-    "@types/jasmine": "~5.1.0",
-    "@types/node": "^18.18.0",
-    "angular-eslint": "18.3.1",
-    "autoprefixer": "^10.4.20",
-    "eslint": "^9.9.1",
-    "jasmine-core": "~5.2.0",
-    "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.2.0",
-    "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.1.0",
-    "ng-packagr": "^18.2.0",
-    "postcss": "^8.4.47",
-    "tailwindcss": "^3.4.13",
-    "typescript": "~5.5.2",
-    "typescript-eslint": "8.2.0"
-  }
+    "name": "notion-workspace",
+    "version": "0.0.0-beta",
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve",
+        "build": "ng build",
+        "watch": "ng build --watch --configuration development",
+        "test": "ng test",
+        "lint": "ng lint",
+        "ngx-notion-cms:lint": "ng lint ngx-notion-cms",
+        "ngx-notion-cms:test": "ng test ngx-notion-cms --no-watch --no-progress --browsers=ChromeHadless",
+        "ngx-notion-cms:build": "ng build ngx-notion-cms",
+        "ngx-notion-cms:publish": "npm run ngx-notion-cms:lint && npm run ngx-notion-cms:build && npm publish dist/ngx-notion-cms/",
+        "serve:ssr:ngx-notion-testbed-app": "node dist/ngx-notion-testbed-app/server/server.mjs"
+    },
+    "private": true,
+    "dependencies": {
+        "@angular/animations": "^18.2.0",
+        "@angular/common": "^18.2.0",
+        "@angular/compiler": "^18.2.0",
+        "@angular/core": "^18.2.0",
+        "@angular/forms": "^18.2.0",
+        "@angular/platform-browser": "^18.2.0",
+        "@angular/platform-browser-dynamic": "^18.2.0",
+        "@angular/platform-server": "^18.2.0",
+        "@angular/router": "^18.2.0",
+        "@angular/ssr": "^18.2.5",
+        "express": "^4.18.2",
+        "ngx-highlightjs": "^12.0.0",
+        "rxjs": "~7.8.0",
+        "tslib": "^2.3.0",
+        "zone.js": "~0.14.10"
+    },
+    "devDependencies": {
+        "@angular-devkit/build-angular": "^18.2.5",
+        "@angular/cli": "^18.2.1",
+        "@angular/compiler-cli": "^18.2.0",
+        "@eslint/js": "^9.11.1",
+        "@types/express": "^4.17.17",
+        "@types/jasmine": "~5.1.0",
+        "@types/node": "^18.18.0",
+        "angular-eslint": "18.3.1",
+        "autoprefixer": "^10.4.20",
+        "eslint": "^9.11.1",
+        "globals": "^15.10.0",
+        "jasmine-core": "~5.2.0",
+        "karma": "~6.4.0",
+        "karma-chrome-launcher": "~3.2.0",
+        "karma-coverage": "~2.2.0",
+        "karma-jasmine": "~5.1.0",
+        "karma-jasmine-html-reporter": "~2.1.0",
+        "ng-packagr": "^18.2.0",
+        "postcss": "^8.4.47",
+        "prettier": "^3.3.3",
+        "prettier-eslint": "^16.3.0",
+        "tailwindcss": "^3.4.13",
+        "typescript": "~5.5.2",
+        "typescript-eslint": "^8.2.0"
+    }
 }
