@@ -1,32 +1,32 @@
 // @ts-check
-const tseslint = require("typescript-eslint");
-const rootConfig = require("../../eslint.config.js");
+const tseslint = require('typescript-eslint');
+const rootConfig = require('../../eslint.config.js');
 
 module.exports = tseslint.config(
-  ...rootConfig,
-  {
-    files: ["**/*.ts"],
-    rules: {
-      "@angular-eslint/directive-selector": [
-        "error",
-        {
-          type: "attribute",
-          prefix: "ngx",
-          style: "camelCase",
+    ...rootConfig,
+    {
+        files: ['**/*.ts'],
+        rules: {
+            '@angular-eslint/directive-selector': [
+                'error',
+                {
+                    type: 'attribute',
+                    prefix: 'ngx',
+                    style: 'camelCase',
+                },
+            ],
+            '@angular-eslint/component-selector': [
+                'error',
+                {
+                    type: 'element',
+                    prefix: 'ngx',
+                    style: 'kebab-case',
+                },
+            ],
         },
-      ],
-      "@angular-eslint/component-selector": [
-        "error",
-        {
-          type: "element",
-          prefix: "ngx",
-          style: "kebab-case",
-        },
-      ],
     },
-  },
-  {
-    files: ["**/*.html"],
-    rules: {},
-  }
+    {
+        files: ['**/*.html'],
+        rules: {},
+    }
 );
