@@ -1,23 +1,30 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+// import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { NgxNotionDatabase } from './ngx-notion-database.component';
+// import { CommonModule } from '@angular/common';
+// import { LayoutContainerComponent } from '../layout-container/layout-container.component';
+// import { NgxNotionDatabaseComponent } from './ngx-notion-database.component';
 
-describe('NgxNotionCmsComponent', () => {
-  let component: NgxNotionDatabase;
-  let fixture: ComponentFixture<NgxNotionDatabase>;
+// describe('NgxNotionDatabaseComponent', () => {
+//     let component: NgxNotionDatabaseComponent;
+//     let fixture: ComponentFixture<NgxNotionDatabaseComponent>;
 
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [NgxNotionDatabase]
-    })
-      .compileComponents();
+//     beforeEach(async () => {
 
-    fixture = TestBed.createComponent(NgxNotionDatabase);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
+//         await TestBed.configureTestingModule({
+//             imports: [NgxNotionDatabaseComponent, CommonModule, LayoutContainerComponent],
+//         }).compileComponents();
+
+//         fixture = TestBed.createComponent(NgxNotionDatabaseComponent);
+//         component = fixture.componentInstance;
+//         fixture.detectChanges();
+//     });
+
+//     it('should create', () => {
+//         TestBed.configureTestingModule({
+//             teardown: { destroyAfterEach: true }, // 👈
+//             // (...)
+//         });
+//         expect(component).toBeTruthy();
+//     });
+// });
