@@ -1,11 +1,9 @@
 import { Directive, ElementRef, inject } from '@angular/core';
 
 @Directive({
-  selector: '[libPostTag]',
-  standalone: true,
+    selector: '[ngxPostTag]',
+    standalone: true,
 })
 export class PostTagDirective {
-  element = inject(ElementRef)
-
-
+    element = inject(ElementRef);
 }
