@@ -1,5 +1,3 @@
-
-
 export interface NotionBlock {
     id: string;
     version: number;
@@ -15,14 +13,13 @@ export interface NotionBlock {
     format?: Format;
 }
 
-
 export interface Format {
     page_icon: string;
     page_cover?: string;
     block_color: string;
     copied_from_pointer: CopiedFromPointer;
-    block_width?: number
-    block_height?: number
+    block_width?: number;
+    block_height?: number;
 }
 
 export interface CopiedFromPointer {
@@ -30,7 +27,6 @@ export interface CopiedFromPointer {
     table: string;
     spaceId: string;
 }
-
 
 export interface Properties {
     title: string[][];
@@ -42,9 +38,8 @@ export interface Properties {
     description?: string[][];
 }
 
-
 export type NotionBlockType =
-    'page'
+    | 'page'
     | 'collection_view_page'
     | 'header'
     | 'sub_header'
@@ -62,9 +57,12 @@ export type NotionBlockType =
     | 'table_of_contents'
     | 'video'
     | 'embed'
-    | 'bookmark'
+    | 'bookmark';
 
-export type MapImageUrl = (image: string, block?: NotionBlock['type']) => string;
+export type MapImageUrl = (
+    image: string,
+    block?: NotionBlock['type']
+) => string;
 
 export interface TableOfContentsEntry {
     id: string;
