@@ -1,2 +1,2 @@
 export * from './block.type';
-export * from './database-item.type'
+export * from './database-item.type';
