@@ -2,9 +2,7 @@
  * Public API Surface of ngx-notion-cms
  */
 
-
-//TODO ngx-notion.service
-export * from './lib/services/notion.service';
 export * from './lib/components/database/ngx-notion-database.component';
-export * from './lib/types'
-export * from './lib/components/notion-page/notion-page.component'
+export * from './lib/components/notion-page/notion-page.component';
+export * from './lib/services/notion.service';
+export * from './lib/types';
